add tests for RegionBasedCustomLimiterRule and split to files
