+++ RateLimiter.Tests/.filenesting.json
@@ -1,0 +1,3 @@
+{
+    "help":"https://go.microsoft.com/fwlink/?linkid=866610"
+}
