+++ RateLimiter.Tests/RateLimiterTest.cs
@@ -7,38 +7,8 @@ using System;
 namespace RateLimiter.Tests;
 
 [TestFixture]
-public class RateLimiterTest
+public partial class RateLimiterTest
 {
-	[Test]
-	public void FixedWindowLimiterRule_should_not_allow_excessive_requests()
-	{
-        var fakeTime = new FakeTimeProvider(startDateTime: DateTimeOffset.UtcNow);
-        var rateLimiter = new RateLimiter();
-		rateLimiter.AddRule("api", new FixedWindowLimiterRule(fakeTime).Configure(new FixedWindowLimiterRuleConfiguration(Window: TimeSpan.FromSeconds(5), Limit: 2)));
-
-		rateLimiter.IsRequestAllowed("api", "user1").Should().BeTrue();
-        rateLimiter.IsRequestAllowed("api", "user1").Should().BeTrue();
-        rateLimiter.IsRequestAllowed("api", "user1").Should().BeFalse();
-
-        fakeTime.Advance(TimeSpan.FromSeconds(5));
-
-        rateLimiter.IsRequestAllowed("api", "user1").Should().BeTrue();
-    }
-
-    [Test]
-    public void FixedWindowLimiterRule_should_have_separate_limits_for_different_users()
-    {
-        var fakeTime = new FakeTimeProvider(startDateTime: DateTimeOffset.UtcNow);
-        var rateLimiter = new RateLimiter();
-        rateLimiter.AddRule("api", new FixedWindowLimiterRule(fakeTime).Configure(new FixedWindowLimiterRuleConfiguration(Window: TimeSpan.FromSeconds(5), Limit: 1)));
-
-        rateLimiter.IsRequestAllowed("api", "user1").Should().BeTrue();
-        rateLimiter.IsRequestAllowed("api", "user1").Should().BeFalse();
-
-        rateLimiter.IsRequestAllowed("api", "user2").Should().BeTrue();
-        rateLimiter.IsRequestAllowed("api", "user2").Should().BeFalse();
-    }
-
     [Test]
     public void Each_resource_should_have_own_rule()
     {
@@ -52,20 +22,6 @@ public class RateLimiterTest
 
         rateLimiter.IsRequestAllowed("api/resource2", "user1").Should().BeTrue();
         rateLimiter.IsRequestAllowed("api/resource2", "user1").Should().BeFalse();
-    }
-
-    [Test]
-    public void SingleRequestPerWindowLimiterRule_should_allow_only_one_request_per_timespan()
-    {
-        var fakeTime = new FakeTimeProvider(startDateTime: DateTimeOffset.UtcNow);
-        var rateLimiter = new RateLimiter();
-        rateLimiter.AddRule("api", new SingleRequestPerWindowLimiterRule(fakeTime).Configure(TimeSpan.FromSeconds(5)));
-
-        rateLimiter.IsRequestAllowed("api", "user1").Should().BeTrue();
-        rateLimiter.IsRequestAllowed("api", "user1").Should().BeFalse();
-
-        rateLimiter.IsRequestAllowed("api", "user2").Should().BeTrue();
-        rateLimiter.IsRequestAllowed("api", "user2").Should().BeFalse();
     }
 
     [Test]
